split classes from generic
split the classes from generic. Ranges and vectors etc. are now separate
but still included by generic (which is only a collection of generic
free functions now)
