+++ Plinth/Sizes.hpp
@@ -1,0 +1,7 @@
+#ifndef PLINTH_SIZES_HPP
+#define PLINTH_SIZES_HPP
+
+#include "Size2.hpp"
+#include "Size3.hpp"
+
+#endif // PLINTH_SIZES_HPP
