+++ Plinth/Vectors.hpp
@@ -1,0 +1,7 @@
+#ifndef PLINTH_VECTORS_HPP
+#define PLINTH_VECTORS_HPP
+
+#include "Vector2.hpp"
+#include "Vector3.hpp"
+
+#endif // PLINTH_VECTORS_HPP
